Move createUseDataHook into own file
